refactor(DI): Write more tests and move tests to correct class
