+++ src/test/java/de/griefed/serverpackcreator/ConfigurationTest.java
@@ -365,4 +365,135 @@ class ConfigurationTest {
         Assertions.assertFalse(result);
         new File("forge-manifest.json").delete();
     }
+
+    @SuppressWarnings("ResultOfMethodCallIgnored")
+    @Test
+    void writeConfigToFileTestFabric() {
+        String modpackDir = "./src/test/resources/fabric_tests";
+        List<String> clientMods = Arrays.asList(
+                "AmbientSounds",
+                "BackTools",
+                "BetterAdvancement",
+                "BetterPing",
+                "cherished",
+                "ClientTweaks",
+                "Controlling",
+                "DefaultOptions",
+                "durability",
+                "DynamicSurroundings",
+                "itemzoom",
+                "jei-professions",
+                "jeiintegration",
+                "JustEnoughResources",
+                "MouseTweaks",
+                "Neat",
+                "OldJavaWarning",
+                "PackMenu",
+                "preciseblockplacing",
+                "SimpleDiscordRichPresence",
+                "SpawnerFix",
+                "TipTheScales",
+                "WorldNameRandomizer"
+        );
+        List<String> copyDirs = Arrays.asList(
+                "config",
+                "mods",
+                "scripts",
+                "seeds",
+                "defaultconfigs"
+        );
+        String javaPath;
+        String autoJavaPath = System.getProperty("java.home").replace("\\", "/") + "/bin/java";
+        if (autoJavaPath.startsWith("C:")) {
+            autoJavaPath = String.format("%s.exe", autoJavaPath);
+        }
+        if (new File("/usr/bin/java").exists()) {
+            javaPath = "/usr/bin/java";
+        } else {
+            javaPath = autoJavaPath;
+        }
+        String minecraftVersion = "1.16.5";
+        String modLoader = "Fabric";
+        String modLoaderVersion = "0.11.3";
+        boolean result = configuration.writeConfigToFile(
+                modpackDir,
+                configuration.buildString(clientMods.toString()),
+                configuration.buildString(copyDirs.toString()),
+                true,
+                javaPath,
+                minecraftVersion,
+                modLoader,
+                modLoaderVersion,
+                true,
+                true,
+                true,
+                true,
+                configuration.getConfigFile(),
+                false
+        );
+        Assertions.assertTrue(result);
+        Assertions.assertTrue(new File("./serverpackcreator.conf").exists());
+        new File("./serverpackcreator.conf").delete();
+    }
+
+    @SuppressWarnings("ResultOfMethodCallIgnored")
+    @Test
+    void testWriteConfigToFileForge() {
+        String modpackDir = "./src/test/resources/forge_tests";
+        List<String> clientMods = Arrays.asList(
+                "AmbientSounds",
+                "BackTools",
+                "BetterAdvancement",
+                "BetterPing",
+                "cherished",
+                "ClientTweaks",
+                "Controlling",
+                "DefaultOptions",
+                "durability",
+                "DynamicSurroundings",
+                "itemzoom",
+                "jei-professions",
+                "jeiintegration",
+                "JustEnoughResources",
+                "MouseTweaks",
+                "Neat",
+                "OldJavaWarning",
+                "PackMenu",
+                "preciseblockplacing",
+                "SimpleDiscordRichPresence",
+                "SpawnerFix",
+                "TipTheScales",
+                "WorldNameRandomizer"
+        );
+        List<String> copyDirs = Arrays.asList(
+                "config",
+                "mods",
+                "scripts",
+                "seeds",
+                "defaultconfigs"
+        );
+        String javaPath = "/use/bin/java";
+        String minecraftVersion = "1.16.5";
+        String modLoader = "Forge";
+        String modLoaderVersion = "36.1.2";
+        boolean result = configuration.writeConfigToFile(
+                modpackDir,
+                configuration.buildString(clientMods.toString()),
+                configuration.buildString(copyDirs.toString()),
+                true,
+                javaPath,
+                minecraftVersion,
+                modLoader,
+                modLoaderVersion,
+                true,
+                true,
+                true,
+                true,
+                configuration.getConfigFile(),
+                false
+        );
+        Assertions.assertTrue(result);
+        Assertions.assertTrue(new File("./serverpackcreator.conf").exists());
+        new File("./serverpackcreator.conf").delete();
+    }
 }
